Refactored parameter parsing into a parameter.rs file
